fixed mobile nav dropdown staying open when clicked off. fixed overflow on about page
