+++ app/components/nav/Navbar.tsx
@@ -1,5 +1,6 @@
 import Link from 'next/link'
-import DarkModeToggle from './DarkModeToggle'
+import DarkModeToggle from '../DarkModeToggle'
+import MobileNavDropdown from './MobileNavDropdown'
 
 const Logo = ({ className }: { className?: string }) => {
   return (
@@ -7,26 +8,7 @@ const Logo = ({ className }: { className?: string }) => {
   )
 }
 
-const HamburgerDropdownIcon = () => {
-  return (
-    <summary className="btn btn-ghost">
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        className="h-5 w-5"
-        fill="none"
-        viewBox="0 0 24 24"
-        stroke="currentColor">
-        <path
-          strokeLinecap="round"
-          strokeLinejoin="round"
-          strokeWidth="2"
-          d="M4 6h16M4 12h8m-8 6h16" />
-      </svg>
-    </summary>
-  )
-}
-
-const navLinks = [
+export const navLinks = [
   {
     label: 'about',
     href: '/about',
@@ -63,14 +45,7 @@ const Navbar = () => {
           </li>
         </ul>
         {/* mobile */}
-        <details className="dropdown dropdown-start md:hidden">
-          <HamburgerDropdownIcon />
-          <ul tabIndex={0} className="dropdown-content menu z-[1] bg-base-200 p-6 rounded-box shadow w-56 gap-2">
-            {navLinks.filter(link => link.mobile).map((link, index) => (
-              <li key={index}><Link href={link.href}>{link.label}</Link></li>
-            ))}
-          </ul>
-        </details>
+        <MobileNavDropdown />
         <Link href="/"><Logo className="md:hidden" /></Link>
         <div className="md:hidden">
           <DarkModeToggle />
